% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{CopyClust}
\alias{CopyClust}
\title{CopyClust Model Prediction}
\usage{
CopyClust(data_input, model_approach = "10C")
}
\arguments{
\item{data_input}{A data frame with sample IDs as rows and 478 model features as columns.}

\item{model_approach}{If equal to "10C", implement 10-class model approach. If equal to "6C", will implement 6-class model approach with binary reclassification.}
}
\value{
A numeric vector of predicted Integrative Cluster label according to model approach.
}
\description{
CopyClust Model Prediction
}
