% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{CopyClust_Model_Prediction}
\alias{CopyClust_Model_Prediction}
\title{CopyClust Model Prediction}
\usage{
CopyClust_Model_Prediction(data_input)
}
\arguments{
\item{data_input}{A data frame with sample IDs as rows and 478 model features as columns.}
}
\value{
A numeric vector of predicted Integrative Cluster label.
}
\description{
CopyClust Model Prediction
}
