% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg18_ranges}
\alias{hg18_ranges}
\title{hg18 Ranges}
\format{
\subsection{\code{hg18_ranges}}{

A 478 by 5 matrix
}
}
\usage{
hg18_ranges
}
\description{
Genomic ranges used as features for the CopyClust function linked to the hg18 reference genome.
Columns: "chrom", "start", "end", "width", "range".
}
\keyword{datasets}
