% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_ranges}
\alias{hg19_ranges}
\title{hg19 Ranges}
\format{
\subsection{\code{hg19_ranges}}{

A 478 by 5 matrix
}
}
\usage{
hg19_ranges
}
\description{
Genomic ranges used as features for the CopyClust function linked to the hg19 reference genome.
Columns: "chrom", "start", "end", "width", "range".
}
\keyword{datasets}
