% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_ranges}
\alias{hg38_ranges}
\title{hg38 Ranges}
\format{
\subsection{\code{hg38_ranges}}{

A 478 by 5 matrix
}
}
\usage{
hg38_ranges
}
\description{
Genomic ranges used as features for the CopyClust function linked to the hg38 reference genome.
Columns: "chrom", "start", "end", "width", "range".
}
\keyword{datasets}
