% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{CopyClust Example Data from TCGA}
\format{
\subsection{\code{test_data}}{

A 1075 by 478 matrix
}
}
\usage{
test_data
}
\description{
Example data from TCGA samples used to test the package
}
\keyword{datasets}
