% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data_raw}
\alias{test_data_raw}
\title{CopyClust Example Data from TCGA Raw}
\format{
\subsection{\code{test_data_raw}}{

<<<<<<< HEAD
A 1075 by 478 matrix
=======
A 6 by 211907 matrix
>>>>>>> 3ad6b6b (added raw test data)
}
}
\usage{
test_data_raw
}
\description{
Example data from TCGA samples used to test the package raw copy number values in DNACopy format
}
\keyword{datasets}
