% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_labels}
\alias{test_labels}
\title{CopyClust Example TCGA Integrative Cluster Labels}
\format{
\subsection{\code{test_labels}}{

A 1075 by 2 matrix
}
}
\usage{
test_labels
}
\description{
Gold standard integrative cluster labels of 1,075 example TCGA samples from 'test_data' determined using combined copy number and gene expression classifier from iC10 package.
}
\keyword{datasets}
