% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{CC_format}
\alias{CC_format}
\title{Format Data for CopyClust Function}
\usage{
CC_format(data_input, reference_genome = "hg18", probes = 1e+05)
}
\arguments{
\item{data_input}{A data frame representing the output of DNAcopy. Six columns: "ID", "chrom", "loc.start", "loc.end", "num.mark", "seg.mean"}

\item{reference_genome}{Parameter for reference genome. Formats the genomic ranges to the appropriate reference genome. Valid inputs are "hg18", "hg19", and "hg38". Default is "hg18".}

\item{probes}{Parameter for number of probes to utilize. Specified number of probes will be used to calculate values for model features. Specified number of probes will be selected from all available probes equally spaced across the genome. Default is 100,000 probes. A greater number of probes decreases the processing speed.}
}
\value{
A data frame with sample IDs as rows and 478 model features as columns that can be used with the CopyClust function.
}
\description{
Formats raw data from DNACopy format into 478 genomic range features required to run the CopyClust() function.
Reference genome (hg18, h19, or hg38) must be specified.
}
