% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data_raw}
\alias{test_data_raw}
\title{CopyClust Example Data from TCGA Raw}
\format{
\subsection{\code{test_data_raw}}{

A 20413 by 6 matrix
}
}
\usage{
test_data_raw
}
\description{
Example raw copy number values in DNACopy format from 100 TCGA samples used to test the CC_format function.
}
\keyword{datasets}
