% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_labels}
\alias{test_labels}
\title{CopyClust Example TCGA IntClust Labels}
\format{
\subsection{\code{test_labels}}{

A 1075 by 2 matrix
}
}
\usage{
test_labels
}
\description{
IntClust labels of example TCGA samples
}
\keyword{datasets}
